[build] E: Make Clippy Happy
