+++ src/pm/process/mod.rs
@@ -8,7 +8,7 @@
 mod capability;
 mod identity;
 mod manager;
-mod process;
+mod state;
 
 //==================================================================================================
 // Exports
