+++ src/pm/process/state/interrupted.rs
@@ -8,8 +8,8 @@
 use crate::{
     hal::arch::ContextInformation,
     pm::{
-        process::process::{
-            state::ProcessState,
+        process::state::{
+            ProcessState,
             RunningProcess,
         },
         thread::InterruptedThread,
