+++ src/pm/process/state/mod.rs
@@ -2,6 +2,16 @@
 // Licensed under the MIT License.
 
 //==================================================================================================
+// Modules
+//==================================================================================================
+
+mod interrupted;
+mod runnable;
+mod running;
+mod suspended;
+mod zombie;
+
+//==================================================================================================
 // Imports
 //==================================================================================================
 
@@ -40,6 +50,63 @@ use ::sys::{
         UserIdentifier,
     },
 };
+
+//==================================================================================================
+// Exports
+//==================================================================================================
+
+pub use interrupted::{
+    InterruptReason,
+    InterruptedProcess,
+};
+pub use runnable::RunnableProcess;
+pub use running::RunningProcess;
+pub use suspended::SleepingProcess;
+pub use zombie::ZombieProcess;
+
+//==================================================================================================
+// ProcessRefMut
+//==================================================================================================
+
+pub enum ProcessRefMut<'a> {
+    Runnable(&'a mut RunnableProcess),
+    Running(&'a mut RunningProcess),
+    Sleeping(&'a mut SleepingProcess),
+    Interrupted(&'a mut InterruptedProcess),
+    Zombie(&'a mut ZombieProcess),
+}
+
+impl<'a> ProcessRefMut<'a> {
+    pub fn state_mut(&mut self) -> &mut ProcessState {
+        match self {
+            ProcessRefMut::Runnable(process) => process.state_mut(),
+            ProcessRefMut::Running(process) => process.state_mut(),
+            ProcessRefMut::Sleeping(process) => process.state_mut(),
+            ProcessRefMut::Interrupted(process) => process.state_mut(),
+            ProcessRefMut::Zombie(process) => process.state_mut(),
+        }
+    }
+}
+
+pub enum ProcessRef<'a> {
+    Runnable(&'a RunnableProcess),
+    Running(&'a RunningProcess),
+    Sleeping(&'a SleepingProcess),
+    Interrupted(&'a InterruptedProcess),
+    Zombie(&'a ZombieProcess),
+}
+
+impl<'a> ProcessRef<'a> {
+    pub fn state(&self) -> &ProcessState {
+        match self {
+            ProcessRef::Runnable(process) => process.state(),
+            ProcessRef::Running(process) => process.state(),
+            ProcessRef::Sleeping(process) => process.state(),
+            ProcessRef::Interrupted(process) => process.state(),
+            ProcessRef::Zombie(process) => process.state(),
+        }
+    }
+}
 
 //==================================================================================================
 // Process
@@ -196,7 +263,7 @@ impl ProcessState {
             None => {
                 let reason: &'static str = "io port not found";
                 error!("remove_pmio(): {:?}", reason);
-                Err(Error::new(ErrorCode::NoSuchEntry, &reason))
+                Err(Error::new(ErrorCode::NoSuchEntry, reason))
             },
         }
     }
@@ -208,7 +275,7 @@ impl ProcessState {
             None => {
                 let reason: &'static str = "io port not found";
                 error!("get_pmio(): {:?}", reason);
-                Err(Error::new(ErrorCode::NoSuchEntry, &reason))
+                Err(Error::new(ErrorCode::NoSuchEntry, reason))
             },
         }
     }
@@ -235,7 +302,7 @@ impl ProcessState {
             None => {
                 let reason: &'static str = "io port not found";
                 error!("get_pmio_mut(): {:?}", reason);
-                Err(Error::new(ErrorCode::NoSuchEntry, &reason))
+                Err(Error::new(ErrorCode::NoSuchEntry, reason))
             },
         }
     }
