+++ src/pm/process/state/running.rs
@@ -8,9 +8,9 @@
 use crate::{
     hal::arch::ContextInformation,
     pm::{
-        process::process::{
-            state::ProcessState,
+        process::state::{
             suspended::SleepingProcess,
+            ProcessState,
             RunnableProcess,
             ZombieProcess,
         },
